+++ app/src/main/java/com/project/simplecoffee/domain/model/Drink.kt
@@ -4,6 +4,6 @@ data class Drink(
     var name: String? = null,
     val image_url: String? = null,
     val price: Double? = null,
-    val stock: Long? = null,
+    val stock: Int? = null,
     val category: List<DrinkCategory>? = null
 ) : Model()
